Update README.md
The Tip Calculator App is a simple yet effective tool designed to help users calculate the tip amount based on the percentage of the total bill and the number of people splitting the bill. The app provides a user-friendly interface built using Kotlin and Jetpack Compose, making it easy to navigate and utilize.